requirements.txt
asgiref==3.6.0
awsebcli==3.20.6
botocore==1.29.99
cement==2.8.2
certifi==2023.5.7
charset-normalizer==2.0.12
colorama==0.4.3
distlib==0.3.5
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.7.1
idna==3.4
jmespath==1.0.1
pathspec==0.10.1
platformdirs==2.5.2
PyJWT==2.7.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
termcolor==1.1.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.16.2
wcwidth==0.1.9
